node:vm: validate the run timeout as a positive uint32 like Node
getRunInContextArgs calls validateUint32(timeout, 'options.timeout', true),
so Node rejects a timeout above 2 ** 32 - 1 with ERR_OUT_OF_RANGE. The run
methods accepted anything up to int64 max. SourceTextModule#evaluate()
already uses validateUint32 in vm.ts.

+++ src/jsc/bindings/NodeVMScript.h
@@ -108,7 +108,7 @@ private:
 class RunningScriptOptions {
 public:
     bool displayErrors = true;
-    std::optional<int64_t> timeout = std::nullopt;
+    std::optional<uint32_t> timeout = std::nullopt;
     bool breakOnSigint = false;
 
     bool fromJS(JSC::JSGlobalObject* globalObject, JSC::VM& vm, JSC::ThrowScope& scope, JSC::JSValue optionsArg);
